writeAlizer: An R Package to Generate Automated Writing Quality and Curriculum-Based Measurement (CBM) Scores.

Package-level documentation for writeAlizer.

Details

Detailed documentation on writeAlizer is available in the GitHub README file and wiki.

The writeAlizer R package (a) imports ReaderBench, Coh-Metrix, and GAMET output files into R, and (b) uses research-developed scoring models to generate predicted writing quality scores or Correct Word Sequences and Correct Minus Incorrect Word Sequences scores from those files.

The writeAlizer package includes functions to do two types of tasks: (1) importing ReaderBench, Coh-Metrix, and/or GAMET output files into R; and (2) generating predicted quality scores using the imported output files. There are also additional functions to help with (3) installation of package dependencies and (4) cache management.

1. Import output files

import_rb

import_coh

import_gamet

import_merge_gamet_rb

2. Generate predicted quality scores

predict_quality

3. Identify necessary packages

model_deps

4. Cache management

wa_cache_dir

wa_cache_clear

See also

Author

Maintainer: Sterett H. Mercer sterett.mercer@ubc.ca (ORCID)